vtpm: Get the supported capabilities of swtpm and swtpm_setup
Call 'swtpm chardev --print-capabilities' to get the supported capabilites
from swtpm. An JSON object is printed by swtpm that we unmarshal and
we pick the 'features' part from it that is an array of strings indicating
what this version of swtpm supports. This option was added in v0.2.
For older versions of swtpm we return an empty array.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>