travis: don't hardcode golang versions
Golang is supporting two minor releases (currently 1.14.x and 1.13.x),
and thus once a new minor release is out, the oldest supported one
becomes unsupported (e.g. when 1.15 is out, 1.13 becomes unsupported).

Instead of hardcoding golang versions, let's specify "stable" and
"oldstable" (which currently equals to 1.14.x and 1.13.x). This way,
we'll be sure we're testing stuff using up to date and supported golang
releases.

This was not possible before due to:
 - https://github.com/travis-ci/gimme/issues/179
 - https://github.com/travis-ci/gimme/issues/185

but apparently it's finally fixed. Hooray!

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>