rootfs: handle github.com/moby/sys/mountinfo deficiency
github.com/moby/sys/mountinfo doesn't fill the Root field of
mountinfo.Info when calling filter functions, meaning that the change in
commit b8bf57281298 ("rootfs: handle nested procfs mounts for MS_MOVE")
regressed the security hardening.

While this is being fixed upstream in [1], just remove this info.Root
optimisation because the ENOENT check is actually sufficient to fix the
original issue.

[1]: https://github.com/moby/sys/pull/50

Reported-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>