[hotfix] seccomp: default to -ENOSYS for SECCOMP_RET_ERRNO
This is a hotfix to make -ENOSYS the default errno if the default action
is SECCOMP_RET_ERRNO. This is necessary because glibc cannot make use of
newer syscalls if we block all unknown syscalls with -EPERM (our old
behaviour).

Unfortunately this is not an ideal solution (syscalls with complicated
rules will now return -ENOSYS rather than -EPERM) but a complete
solution will require far more work -- most likely a reimplementation of
libseccomp to allow us to create custom BPF filters ourselves, or large
changes to libseccomp to better accommodate our requirements -- and thus
this hotfix was written to solve the immediate problem while we work on
a more complete solution.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>