seccomp: fix the inconsistent check of seccomp mode
When it reads from /proc/self/status,
runc cannot check the seccomp mode until seccomp is actually set up.
And the current code just checks the existence of key which is "Seccomp".
On the other hand, Prctl() shows the supported mode corretly.
So I just removed the code which parses /proc/self/status.

Signed-off-by: Kenta Tada <Kenta.Tada@sony.com>