libct/cg/sd: fix dbus connection leak
When running libcontainer/integration/TestSystemdFreeze 2000 times,
I got "too many open files" error. Indeed, systemd cgroup manager
never closes its dbus connection.

This was not a problem before because we only had a single connection
for the whole runtime. Now it is per manager instance, so we leak a
connection (apparently two sockets) per cgroup manager instance.

The fix is to

1. Close the connection from cgroup manager's Destroy (assuming the
   cgroup manager won't be used after Destroy -- in case it will be,
   a new connection will be initiated). This helps for cases when
   the container (and its cgroup) are actually removed.

2. Set a finalizer in newDbusConnManager(), so once it is garbage
   collected, the closeConnection() method is called. This helps for
   cases of a long-running runtime (such as kubernetes) which uses
   libcontainer's cgroup manager to create a cgroup.

A test case is to be added in the next commit.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>