freezer: add delay after freeze
I hate to keep adding those kludges, but lately TestFreeze (and
TestSystemdFreeze) from libcontainer/integration fails a lot. The
failure comes and goes, and is probably this is caused by a slow host
allocated for the test, and a slow VM on top of it.

To remediate, add a small sleep on every 25th iteration in between
asking the kernel to freeze and checking its status.

In the worst case scenario (failure to freeze) this adds 0.4 μs
to the duration of the call (nothing compared to that sleep after
the temporary thaw).

It is hard to measure how this affects CI but (with added debug prints)
on a histogram of number of retries I saw peaks at and after numbers
25, 50, 75 etc. meaning this works.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>