runc -v: set default for, always show main.version
Apparently not everyone compiles runc via the provided Makefile. For
example, one can just run "go build", in which case Version variable
is left empty, which leads to:

	$ ./runc -v
	runc version spec: 1.0.2-dev
	go: go1.16.3

Surely, the main problem here is runc was built in a wrong way, but the
second problem is such output is very confusing -- it may seem that we
have runc 1.0.2.

To solve, make sure to _always_ add version (even if empty), and set the
default to "unknown".

NOTE this does not change anything in case runc is compiled via the
Makefile.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>