chown cgroup to uid 0 in container namespace
Delegating cgroups to the container enables more complex workloads
including systemd-based workloads.  When using the systemd cgroupV2
cgroup manager, chown the container's cgroup directory and the
cgroup.procs, cgroup.subtree_control and cgroup.threads files to the
host UID corresponding to UID 0 in the container.

As a result of this change, systemd workloads can run in isolated
user namespaces on OpenShift.

Signed-off-by: Fraser Tweedale <ftweedal@redhat.com>